host: Rename imports.h to import_fn.h
