+++ src/emulator/host/CMakeLists.txt
@@ -3,7 +3,7 @@ host
 STATIC
 include/host/config.h
 include/host/functions.h
-include/host/imports.h
+include/host/import_fn.h
 include/host/sfo.h
 include/host/state.h
 include/host/version.h
