+++ src/emulator/module/include/module/bridge.h
@@ -22,7 +22,7 @@
 #include "vargs.h"
 #include "write_return_value.h"
 
-#include <host/imports.h>
+#include <host/import_fn.h>
 #include <host/state.h>
 
 #include <microprofile.h>
